+++ phison_realestate_mobile/lib/pages/profile/view/profile_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:phison_realestate_mobile/pages/profile/view/change_language_bottom_sheet.dart';
-import 'package:phison_realestate_mobile/presentation/constants/app_assets_constant.dart';
+import 'package:phison_realestate_mobile/shared/constants/app_assets_constant.dart';
 import 'package:phison_realestate_mobile/shared/widgets/phison_app_bar.dart';
 
 import '../../app/bloc/bloc/app_bloc.dart';
