+++ phison_realestate_mobile/lib/shared/widgets/bottom_navigation_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:phison_realestate_mobile/presentation/constants/app_assets_constant.dart';
+import 'package:phison_realestate_mobile/shared/constants/app_assets_constant.dart';
 
 class PhisonBottomBar extends StatelessWidget {
   const PhisonBottomBar({
